+++ classes/Markdown.php
@@ -35,13 +35,13 @@ class Markdown
     /**
      * Instantiate Markdown API
      *
-     * @param Twig   $twig   Twig-instance
-     * @param Source $source Source-instance
+     * @param Twig   $Twig   Twig-instance
+     * @param Source $Source Source-instance
      */
-    public function __construct(Twig $twig, Source $source)
+    public function __construct(Twig $Twig, Source $Source)
     {
-        $this->twig = $twig;
-        $this->source = $source;
+        $this->Twig = $Twig;
+        $this->Source = $Source;
     }
 
     /**
@@ -152,13 +152,13 @@ class Markdown
         foreach ($matches as $match) {
             $attrs = array();
             $attrs['src'] = $match['file'] . '.' . $match['ext'];
-            $source = $this->source->render($attrs['src'], '', $match['mediaActions']);
-            $attrs['src'] = $source['src'];
-            if (isset($source['filename']) && !empty($source['filename'])) {
-                $filename = $source['filename'];
+            $Source = $this->Source->render($attrs['src'], '', $match['mediaActions']);
+            $attrs['src'] = $Source['src'];
+            if (isset($Source['filename']) && !empty($Source['filename'])) {
+                $filename = $Source['filename'];
             }
-            if (isset($source['page']) && $source['page'] instanceof Page) {
-                $page = $source['page'];
+            if (isset($Source['page']) && $Source['page'] instanceof Page) {
+                $page = $Source['page'];
             }
             $attrs['alt'] = (isset($match['alt']) ? $match['alt'] : '');
             $attrs['title'] = (isset($match['title']) ? $match['title'] : '');
@@ -183,7 +183,7 @@ class Markdown
                 $match[0] = str_replace($match['url'], '', $match[0]);
                 $url = trim($match['url'], '_');
             }
-            $replace = $this->twig->processTemplate(
+            $replace = $this->Twig->processTemplate(
                 'partials/figure.html.twig',
                 [
                     'attrs' => $attrs,